+++ cmd/crypto_http.go
@@ -14,90 +14,15 @@
 package cmd
 
 import (
-	"context"
-	"encoding/base64"
-
-	"github.com/pulumi/pulumi/pkg/util/contract"
+	"github.com/pulumi/pulumi/pkg/secrets/service"
 
 	"github.com/pulumi/pulumi/pkg/backend/httpstate"
-	"github.com/pulumi/pulumi/pkg/backend/httpstate/client"
-	"github.com/pulumi/pulumi/pkg/resource/config"
 	"github.com/pulumi/pulumi/pkg/secrets"
 )
 
-// cloudCrypter is an encrypter/decrypter that uses the Pulumi cloud to encrypt/decrypt a stack's secrets.
-type cloudCrypter struct {
-	client *client.Client
-	stack  client.StackIdentifier
-}
-
-func newCloudCrypter(client *client.Client, stack client.StackIdentifier) config.Crypter {
-	return &cloudCrypter{client: client, stack: stack}
-}
-
-func (c *cloudCrypter) EncryptValue(plaintext string) (string, error) {
-	ciphertext, err := c.client.EncryptValue(context.Background(), c.stack, []byte(plaintext))
-	if err != nil {
-		return "", err
-	}
-	return base64.StdEncoding.EncodeToString(ciphertext), nil
-}
-
-func (c *cloudCrypter) DecryptValue(cipherstring string) (string, error) {
-	ciphertext, err := base64.StdEncoding.DecodeString(cipherstring)
-	if err != nil {
-		return "", err
-	}
-	plaintext, err := c.client.DecryptValue(context.Background(), c.stack, ciphertext)
-	if err != nil {
-		return "", err
-	}
-	return string(plaintext), nil
-}
-
-type cloudSecretsManagerState struct {
-	URL     string `json:"url,omitempty"`
-	Owner   string `json:"owner"`
-	Project string `json:"project"`
-	Stack   string `json:"stack"`
-}
-
-var _ secrets.Manager = &cloudSecretsManager{}
-
-type cloudSecretsManager struct {
-	state   cloudSecretsManagerState
-	crypter config.Crypter
-}
-
-func (sm *cloudSecretsManager) Type() string {
-	return "pulumi"
-}
-
-func (sm *cloudSecretsManager) State() interface{} {
-	return sm.state
-}
-
-func (sm *cloudSecretsManager) Decrypter() (config.Decrypter, error) {
-	contract.Assert(sm.crypter != nil)
-	return sm.crypter, nil
-}
-
-func (sm *cloudSecretsManager) Encrypter() (config.Encrypter, error) {
-	contract.Assert(sm.crypter != nil)
-	return sm.crypter, nil
-}
-
-func newCloudSecretsManager(s httpstate.Stack) secrets.Manager {
-	b := s.Backend().(httpstate.Backend)
+func newCloudSecretsManager(s httpstate.Stack) (secrets.Manager, error) {
+	client := s.Backend().(httpstate.Backend).Client()
 	id := s.StackIdentifier()
 
-	return &cloudSecretsManager{
-		state: cloudSecretsManagerState{
-			URL:     b.CloudURL(),
-			Owner:   id.Owner,
-			Project: id.Project,
-			Stack:   id.Stack,
-		},
-		crypter: newCloudCrypter(b.Client(), id),
-	}
+	return service.NewCloudSecretsManager(client, id)
 }
